// frontend/src/api/generated/model/task.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Gantry Board API
 * AI Agent Orchestration Kanban Board API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  /** @nullable */
  assigned_to?: string | null;
  created_at: string;
  /** @nullable */
  description?: string | null;
  id: string;
  /** @nullable */
  parent_id?: string | null;
  position: number;
  priority: TaskPriority;
  project_id: string;
  status: TaskStatus;
  title: string;
  updated_at: string;
}